Apply fixes for Archlinux release
Archlinux requires some special casing to make this program work, and
this patch should fix those issues.

One of the main issues is that the `PKGBUILD` used by Archlinux requires
installation to happen in `/usr/bin`, but also makes binaries available
in `/bin`. This results in `lonestar` being ran from `/bin`, but
installation assumed it would be ran from `/usr/bin`, and this broke the
paths it uses to find its library files.

To solve this, the configuration file loading has been updated to look
at static file locations in `/etc` and `/usr/local/etc`. These are the
only common places I know off that are used, more can be added if
necessary for other distributions.

Additionally, the `etc` directory ended up being at `/usr/etc`, which
isn't common in distributions. The `Makefile` has been updated to allow
setting a `$ETCDIR` variable to change the directory for configuration
files. It defaults to `/etc`.