Update the post-install notice
It seems users were forgetting to run the `init` part of LoneStar after
installation, so it would probably help to remind them of it. This
particular user didn't read the readme to begin with, so with the
addition of the notice users like this one should be able to get going
without troubles as well.